Resolve incorrect success detection
This was a good one. The final step of the pipeline was failing
sometimes even when the matching phrase was present in the `git
push` output. It turns out the failing exit code was coming from
git and not the right most command in the pipe, grep.

This is because common.sh from bitbucket enables pipefail.

The reason that `git push` was failing was that we were using the
`-q` flag on grep. This was closing the pipe as soon as a search
term was matched. Git was still trying to write to this pipe and
was terminated with a SIGPIPE.